Update WoWPro_Trade.lua
commented out some code, maybe i'm missing something here but it wasn't working, the index was never updating so when I went to look here I see the problem is probably with that for loop at line 312 doesn't write the index properly. But it seems redundant, you are doing the same thing above but in a different way, so why not just update it from the start?