Update and rename 15_18_Redridge.lua to 15_17_Redridge.lua
only significant changes are taking [The Price of Shoes] when you run to get the fp at the end of elwynn guide, so turning in the followup here and being able to do the Underbelly Scales while doing redridge goulosh, since you can't help but kill the black dragon whelps anyway, may as well get something from it. Removed Elmore's Task/Stormpike Delivery since they are in Elwynn Guide. Changed level to 15-17, there is just not enough xp in this guide to get to 18 without a grind.