Update 20_21_Redridge.lua
moved h to stormwind, as there was no H step in guide and next guide opens with setting Hearth to Darkshire, so there was no need for it to be set in Lakeshire, so may as well have it somewhere useful. then, added H step to go turn in quest in Northshire.