Update 21_30_Hendo_HordeChapter1.lua
I found a quest chain in Thousand Needles that starts in Ashenvale. It actually starts/ends in Thunder Bluff, but the objective is done in Ashenvale. I'm adding it to this guide because it makes more sense to do it now rather than later.

I'm also lumping all of my changes into this one and closing my other push requests.