Fix moving tags when deleting/adding steps
tinsert and tremove both depend on contiguous indexed tables to work
properly. Since some of the key tables are not, the indices for these
tables would not get updated properly. This resulted in some tags being
moved to adjacent steps.

These function calls have been replaced with loops that emulate those
functions, and customized for the particularities of this system.