Combat update fix
So this makes it workable.

I wasn't able to figure out what Ludo wanted to so this will simply overlay an invisible secureactionbutton over the insecure one if you're out of combat and the insecureactionbutton is visible and shown.

I'm sure there are cleaner ways to do it but everything I tried failed except this.