Author updates
* '/wp devcoords' to pop up a small edit box that has the coords to easily allow you to copy and paste into guides.

* updated recorder to have the old way "Advanced Mode" and the new default is standard mode with simplified creation and choices. Not sure ANYONE uses the step editor or add/subtract ect so that is left for advanced mode.

Right now it is a manual switch in the code, want to discuss with Ludo about adding a checkbox or something to enable/disable if deemed necessary.

Added a DATE tag that takes the epoch time relative to GMT, this is specifically to gate steps behind an opening date, like the prepatch event 2nd half of the guide which isn't available until the 2nd week.

Also found and fixed a bug in wowpro.lua that could throw an error if you load a guide with no steps (which can happen in recorder).