Report Issue on Right Click frame menu
Added ability to right click on the step frame to choose "Report Issue". Also had it report the full Guide step. (Ludo is there a better way to do this?)

Also found a bug in the errorlog window where the title wasn't showing up. Fixed that while I was there.