Recorder, Class, Race and _G fixes
The issue with new alliance guides that caused an error on reload

Title will now show the guide name instead of Stormwind or Orgrimmar in recorder.

Fixed the case issue in race and class comparison logic.

Removed the _G. from native lua functions and put them up on the ignore list at the top of the file Gethe pointed out to me.