Maldraxxus and bugs
Only 2 Maldraxxus story chains left, nearly done.

Fixed an issue with EAB tag

Added the flip to the IZ tag I asked for a year ago :P

Added flip to BUFF tag (was going to use it but found a better way) figured leave it in if someone needs it later.

Misc guide edits discovered by the wife.