Frames, themes, LSM, fauxjumpbutton
Added the fauxjumpbutton logic to match that of use target and eab.

added LibSharedMedia = LSM as a local in config, put in custom border

Added a theme in the default profile list. Testing to see how this works, if others have some premade themes to try for.

While playing around with the frames for theming I found some misconfiguations, had to add some if/else on some of the frames to have my new border work.  Will clean that up better.