Item and EAB buttons
Gave |U| tags the same treatment as EAB to hide and show the button when applicable.

updated EAB to check the API if  the  EAB should show instead of checking if the EAB frame is visible in case there are addons that auto hide the EAB frame.

Since I had to run through BFA intro to test EAB, found a few changes.

Since I had to run through DK intro to test U, found a change.