IsInChromieTime
Added check to level tag if you are in chromie time or not. Used by LVL|xx;CT| where xx is the level restriction without Chromie Time.

Updated Legion guides with this, also added the Legion content set for order halls to ensure they are set to Legion.

Adjusted nextguide for Revendreth and and Covenant so they properly goto the next guide.

Updated Goblin quest discussed in Author chat.