WotLK Argent Tourney dailies
a duplicate file based on Ludo's ArgTour in the Alliance Dailies. This file is that one, PLUS Horde quests added (about halfway) tot he file, to make this a Neutral guide instead.  Based on my looking I did not see one specifically for Horde, though there was some Horde Faction in this one.   I understand Discord user Nzall was asking for one, and he was also looking at writing one.  This is a start and or one to 'merge' with his guide if he does one.