Profession complete rewrite
okay, here it is.. finally my rewrite of the profession guides. ALL crafting Profs ShadowLands guides are updated to new format. I will be slowly converting ALL the retail guides to the new format. BOTH the old and new guides are in the guides. To test the new format, please select the SL guide that has 'new' in the names.