Update VAN_BCC_10_19_Ghostlands.lua
LOL... I send the Warlocks to Orgrimmar to pickup the next quest in their class chain without sending them into UC to turn in the first step. Couldn't figure out why I couldn't pick it up. Appears that if it is a LEAD quest, you have to turn it in first or abandon it to continue.

- Added travel steps to complete `Hearts of the Pure|QID|1476|` C steps.

Still WIP
Need to add travel step to Hammerfall to get FP and fly back to UC.