Update VAN_BCC_41_51_HordeChapter3.lua
I oopsed... I was thinking of Blasted Lands instead of Badlands. Because we Skipped questing in Arathi Highlands, we don't have the FPs for that area. I'm working on a work-around using a short quest chain to get ported to Gnomeregan and walking across Dun Morogh to Loch Modan.

I have stopped at Dun Morogh for the moment.