DK Intro Guide Fixes (#2066)
Fixed the issues with the intro guide, next guide is set properly to hellfire, guide is neutral. All names/nicknames were specified appropriately and zone matched that in ZoneData. Also renamed file to match that of the naming convention previously used.

Updated leveling addon to recognize starter guides in Wrath expansion as well as the level 55 Death Knights.