EAB and Functionalize fix (#2178)
Changes in DF require secured buttons to have properly registered clicks and can''t inherit from another button.

Functionalize was spitting out zonedata out of order with mapIDs mixed up, got rid of the for pairs for a crude 1-2500 loop. Ludo might want to revisit this to see if there is a cleaner way to do it.

Will need to remember to adjust this if mapIDs grow higher than 2500.