DF campaign and Thald updates
Fixed the ACTIVE on the gather FP step. Copy pasted and never changed.

Removed the U tag from Jump Start in campaign as there is no item to use.

Moving on - Rank 2 and 3 you are already on top of the tower.

The Awaited Egg-splosion split QO 5 and 6 becuz they are not active at the same time and it had a ? on the window.

added a Portal step to The Sweet Taste of Victory.