Cata Header correction
All of the Cata Daily guides have been updated to fix "WoWPro/WoWPro.lua:1199: attempt to index local 'guide' (a string value)" error.

"Cata" needs to be added as the last variable (It's normally '4')

The Achieve, Profession, and World Event guides will need this done as well. The headers are not modernized with the correct License.