DevZone Command
Created a command to display a Player's coordinates with the ZoneID in a format that can be used as copy and paste into any guide.

It works essentially the same way as the devcoords but with the zoneID as well.

I also cleaned up the chat message for the Where command. The 2nd zone text is unnecessary and I removed the quotes.