Disable TomTom warning
Requested by GrimMorningstar on Curseforge.

Updated the coding to disable the TomTom warning after they have acknowledged the warning at least once. They must click ok for it to work. The timeout feature will not do it.

This will only apply to Retail, as the others do not have the in game navigation and therefor, need TomTom.